Refactor html paste internals
