+++ src/plugins/transfer/htmlTransfer.ts
@@ -3,12 +3,21 @@ import type { Editor } from "../../editor.js";
 import {
   getDOMSelection,
   getSelectionRangeInEditor,
-  domToFragment,
+  TOKEN_BLOCK,
+  TOKEN_TEXT,
+  TOKEN_SOFT_BREAK,
+  TOKEN_VOID,
 } from "../../dom/index.js";
 import { isCommentNode } from "../../dom/utils.js";
-import type { DocNode, InferInlineNode, TextNode } from "../../doc/types.js";
+import type {
+  BlockNode,
+  DocNode,
+  InferInlineNode,
+  InlineNode,
+  TextNode,
+} from "../../doc/types.js";
 import type { PasteHook } from "../../editor.js";
-import type { Parser } from "../../dom/parser.js";
+import type { Parser, TokenType } from "../../dom/parser.js";
 
 type HtmlSerializers<T extends DocNode> = Partial<{
   [key in keyof HTMLElementTagNameMap]: (
@@ -25,6 +34,19 @@ export const htmlPaste = <T extends DocNode>(
   parse: Parser,
   serializers: HtmlSerializers<T>,
 ): PasteHook => {
+  const serializeText = serializers["text"];
+  const serializeVoid = (n: Element) => {
+    const s = serializers[n.tagName.toLowerCase() as keyof typeof serializers];
+    if (s) {
+      const node = s(n as any);
+      if (node) {
+        return node;
+      }
+    }
+
+    return;
+  };
+
   return (dataTransfer) => {
     const html = dataTransfer.getData("text/html");
     if (html) {
@@ -45,18 +67,62 @@ export const htmlPaste = <T extends DocNode>(
       }
 
       // TODO customizable dom to standard schema and validate
-      return domToFragment(dom, parse, serializers["text"], (n) => {
-        const s =
-          serializers[n.tagName.toLowerCase() as keyof typeof serializers];
-        if (s) {
-          const node = s(n as any);
-          if (node) {
-            return node;
+      return parse(({ _next: next, _domNode: domNode }) => {
+        let type: TokenType | void;
+        let row: InlineNode[] | null = null;
+        let text = "";
+        let hasContent = false;
+
+        const rows: BlockNode[] = [];
+
+        const completeText = () => {
+          if (text) {
+            if (!row) {
+              row = [];
+            }
+            row.push(serializeText(text));
+            text = "";
+          }
+        };
+        const completeRow = () => {
+          completeText();
+          if (!row && hasContent) {
+            row = [];
+          }
+          if (row) {
+            rows.push({ children: row });
+          }
+          row = null;
+          hasContent = false;
+        };
+
+        while ((type = next())) {
+          if (type === TOKEN_BLOCK) {
+            completeRow();
+          } else {
+            hasContent = true;
+
+            if (type === TOKEN_TEXT) {
+              text += domNode<typeof type>().data;
+            } else if (type === TOKEN_VOID) {
+              completeText();
+              const docNode = serializeVoid(domNode<typeof type>());
+              if (docNode) {
+                row!.push(docNode);
+              }
+            } else if (type === TOKEN_SOFT_BREAK) {
+              completeRow();
+            }
           }
         }
+        completeRow();
 
-        return;
-      });
+        if (!rows.length) {
+          rows.push({ children: [] });
+        }
+
+        return rows;
+      }, dom);
     }
     return null;
   };
